+++ mirage/factories/contributor.ts
@@ -1,4 +1,5 @@
-import { association, Collection, Factory, faker, trait, Trait } from 'ember-cli-mirage';
+import { association, Collection, Factory, trait, Trait } from 'ember-cli-mirage';
+import faker from 'faker';
 
 import Contributor from 'ember-osf-web/models/contributor';
 import { Permission } from 'ember-osf-web/models/osf-model';
@@ -9,11 +10,12 @@ interface ContributorTraits {
 }
 
 export default Factory.extend<Contributor & ContributorTraits>({
-    permission: faker.list.cycle(...Object.values(Permission)),
-    bibliographic: true,
-    unregisteredContributor() {
-        return faker.random.number(5) ? undefined : faker.name.firstName();
+    permission(i: number) {
+        const permissions = Object.values(Permission);
+        return permissions[i % permissions.length];
     },
+    bibliographic: true,
+    unregisteredContributor: undefined,
     index(i: number) {
         return i;
     },