+++ mirage/factories/registration-provider.ts
@@ -1,4 +1,5 @@
-import { Factory, faker, trait, Trait } from 'ember-cli-mirage';
+import { Factory, trait, Trait } from 'ember-cli-mirage';
+import faker from 'faker';
 
 import RegistrationProvider from 'ember-osf-web/models/registration-provider';
 
@@ -16,6 +17,8 @@ export interface MirageRegistrationProvider extends RegistrationProvider {
 
 export interface RegistrationProviderTraits {
     withBrand: Trait;
+    withSchemas: Trait;
+    submissionsNotAllowed: Trait;
 }
 
 export default Factory.extend<MirageRegistrationProvider & RegistrationProviderTraits>({
@@ -50,6 +53,16 @@ export default Factory.extend<MirageRegistrationProvider & RegistrationProviderT
     withBrand: trait<RegistrationProvider>({
         afterCreate(provider, server) {
             provider.update({ brand: server.create('brand') });
+        },
+    }),
+    withSchemas: trait<RegistrationProvider>({
+        afterCreate(provider, server) {
+            provider.update({ schemas: [server.schema.registrationSchemas.find('testSchema')] });
+        },
+    }),
+    submissionsNotAllowed: trait<RegistrationProvider>({
+        afterCreate(provider) {
+            provider.update({ allowSubmissions: false });
         },
     }),
 });