+++ tests/acceptance/guid-file/file-detail-test.ts
@@ -8,7 +8,7 @@ import {
     triggerKeyEvent,
     visit,
 } from '@ember/test-helpers';
-import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
+import { setupMirage } from 'ember-cli-mirage/test-support';
 import config from 'ember-get-config';
 import { percySnapshot } from 'ember-percy';
 import moment from 'moment';