+++ tests/acceptance/settings/account/default-region-test.ts
@@ -1,5 +1,5 @@
 import { visit } from '@ember/test-helpers';
-import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
+import { setupMirage } from 'ember-cli-mirage/test-support';
 import { selectChoose } from 'ember-power-select/test-support';
 import { module, test } from 'qunit';