+++ tests/acceptance/settings/developer-apps-page-test.ts
@@ -1,5 +1,5 @@
 import { currentRouteName, currentURL, fillIn, visit, waitFor } from '@ember/test-helpers';
-import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
+import { setupMirage } from 'ember-cli-mirage/test-support';
 import { percySnapshot } from 'ember-percy';
 import { module, test } from 'qunit';