+++ tests/engines/registries/acceptance/overview/topbar-test.ts
@@ -1,6 +1,6 @@
 import { triggerEvent } from '@ember/test-helpers';
 import { ModelInstance } from 'ember-cli-mirage';
-import setupMirage from 'ember-cli-mirage/test-support/setup-mirage';
+import { setupMirage } from 'ember-cli-mirage/test-support';
 import { t } from 'ember-intl/test-support';
 import moment from 'moment';
 import { module, test } from 'qunit';