+++ types/ember-cli-mirage/index.d.ts
@@ -3,9 +3,8 @@ import MirageModelRegistry from 'ember-cli-mirage/types/registries/model';
 import MirageSchemaRegistry from 'ember-cli-mirage/types/registries/schema';
 import DS from 'ember-data';
 import EmberDataModelRegistry from 'ember-data/types/registries/model';
+import { BelongsTo } from 'miragejs/-types';
 import { Document } from 'osf-api';
-
-export { default as faker } from 'faker';
 
 declare global {
     // eslint-disable-next-line no-redeclare
@@ -63,7 +62,6 @@ interface ModelInstanceShared<T> {
 }
 
 export function hasMany(model: string): void;
-export function belongsTo(model: string): void;
 
 export type ModelInstance<T = AnyAttrs> = ModelInstanceShared<T> & ModelInstanceAttrs<T>;
 
@@ -261,9 +259,10 @@ export function trait<
 // function association(...traits: string[], overrides?: { [key: string]: any }): any;
 
 export function association(...args: any[]): any;
+export { belongsTo } from 'miragejs';
 
 export type FactoryAttrs<T> = {
-    [P in keyof T]?: T[P] | ((index: number) => T[P]);
+    [P in keyof T]?: T[P] | BelongsTo | ((index: number) => T[P]);
 } & {
     afterCreate?(newObj: ModelInstance<T>, server: Server): void;
 };